requirements.txt
# Bird Classification - Requirements
# Install with: pip install -r requirements.txt

# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.2.0

# Training utilities
tqdm>=4.65.0
tensorboard>=2.12.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development (optional)
ipython>=8.0.0
jupyter>=1.0.0
